frontend/requirements.txt
Flask
mysql-connector
isbnlib
jinja2